Add files via upload
When I was start this project, I have no idea about the web technologies. I start is with the basic HTML and CSS . After some time I made the home page by it's help and continuously try to make it's front end beautiful. And simoultaneously leaning PHP for it's back end development. With the learning of PHP I try to integrate it with the web pages. After a time I learn the MySQL database connection with PHP and I implemented it into this project. After this I try to learn fetching data from the database, and it was work finely. Then I gaining the knowledge about database operation like searching ,insert , update, delete values from the database. And try it to implement through PHP, and it was succesful. Suddenly one day I have seen my virtual server is not working properly, so I decided to reinstall my server. Before that I backup my data from the database and when I try to backup my pages from the working directories , I noticed some files are missing . I was one of the horrible situation in my life, because my 7 to 8 mounth hard worked project more than fifty percent was spoiled. I was really depressed . I have decided that I will make this project again. Now my project is completed within the half of the previous work time, because of my previous experience help me a lot to develop this project. I made a educational institute's website with help of PHP. And there I have tried to add all necessary options for the all respected users.